Refactor HomeController and update configs
Removed unused Azure namespaces and references in HomeController.cs, indicating a possible shift from certain Azure services or a cleanup. Additionally, the direct assignment to ViewBag.sql from cosmosResult was removed, suggesting changes in data handling for views.

In appsettings.json, updated the ConnectionStrings for CosmosDB and StorageAccount with actual values, replacing placeholders. This is crucial for the proper connection to Azure services. Also, enhanced the clarity of settings by renaming keys to include 'Name' and updating their values to specific names, indicating a move towards a new storage and database setup.